# build: replace prettier with biome