--- a/.prettierignore
+++ b/.prettierignore
@@ -1,3 +1,0 @@
-# Avoid yarn and prettier fighting over formatting
-package.json
-**/*.trace.json