--- a/build-utils/last-built-plugin.ts
+++ b/build-utils/last-built-plugin.ts
@@ -1,6 +1,4 @@
 /* eslint-env node */
-/* eslint import/no-nodejs-modules:0 */
-
 import fs from 'fs';
 import path from 'path';