--- a/scripts/extract-ios-device-names.ts
+++ b/scripts/extract-ios-device-names.ts
@@ -1,4 +1,3 @@
-/* eslint-disable import/no-nodejs-modules */
 /* eslint-env node */
 import fs from 'fs';
 import path from 'path';