--- a/scripts/test.js
+++ b/scripts/test.js
@@ -1,6 +1,4 @@
 /* global process */
-/* eslint import/no-nodejs-modules:0 */
-
 // Do this as the first thing so that any code reading it knows the right env.
 // process.env.BABEL_ENV = 'test';
 process.env.NODE_ENV = 'test';