--- a/static/app/api.tsx
+++ b/static/app/api.tsx
@@ -633,7 +633,6 @@ export class Client {
         }
       )
       .catch(error => {
-        // eslint-disable-next-line no-console
         console.error(error);
         Sentry.captureException(error);
       });