--- a/static/app/components/button.tsx
+++ b/static/app/components/button.tsx
@@ -240,7 +240,7 @@ function BaseButton({
     }
 
     return () => {
-      // eslint-disable-next-line no-console
+      // biome-ignore lint/suspicious/noConsoleLog: Disable console
       console.log('buttonAnalyticsEvent', {
         eventKey: analyticsEventKey,
         eventName: analyticsEventName,
@@ -430,8 +430,9 @@ const getColors = ({
       color: ${color};
     }
 
-    ${size !== 'zero' &&
-    `
+    ${
+      size !== 'zero' &&
+      `
       &:hover,
       &:active,
       &[aria-expanded="true"] {
@@ -443,7 +444,8 @@ const getColors = ({
         color: ${colorActive || color};
         border-color: ${borderActive};
       }
-    `}
+    `
+    }
 
     &.focus-visible {
       ${getFocusState()}
@@ -596,7 +598,6 @@ export {
   BaseButtonProps,
   LinkButton,
   LinkButtonProps,
-
   // Also export these styled components so we can use them as selectors
   StyledButton,
   ButtonLabel,