--- a/static/app/components/codeSnippet.tsx
+++ b/static/app/components/codeSnippet.tsx
@@ -108,8 +108,8 @@ export function CodeSnippet({
     tooltipState === 'copy'
       ? t('Copy')
       : tooltipState === 'copied'
-      ? t('Copied')
-      : t('Unable to copy');
+        ? t('Copied')
+        : t('Unable to copy');
 
   return (
     <Wrapper