--- a/static/app/components/deprecatedAsyncComponent.tsx
+++ b/static/app/components/deprecatedAsyncComponent.tsx
@@ -47,7 +47,6 @@ function wrapErrorHandling<T extends any[], U>(
     try {
       return fn(...args);
     } catch (error) {
-      // eslint-disable-next-line no-console
       console.error(error);
       window.setTimeout(() => {
         throw error;
@@ -105,13 +104,13 @@ class DeprecatedAsyncComponent<
     const currentLocation = isLocationInProps
       ? this.props.location
       : isRouterInContext
-      ? this.context.router.location
-      : null;
+        ? this.context.router.location
+        : null;
     const prevLocation = isLocationInProps
       ? prevProps.location
       : isRouterInContext
-      ? prevContext.router.location
-      : null;
+        ? prevContext.router.location
+        : null;
 
     if (!(currentLocation && prevLocation)) {
       return;
@@ -452,8 +451,8 @@ class DeprecatedAsyncComponent<
     return this.shouldRenderLoading
       ? this.renderLoading()
       : this.state.error
-      ? this.renderError()
-      : this.renderBody();
+        ? this.renderError()
+        : this.renderBody();
   }
 
   /**