--- a/static/app/components/events/eventEntry.tsx
+++ b/static/app/components/events/eventEntry.tsx
@@ -147,7 +147,6 @@ function EventEntryContent({
     // this should not happen
     default:
       if (window.console) {
-        // eslint-disable-next-line no-console
         console.error?.('Unregistered interface: ' + (entry as any).type);
       }
       return null;