… .pre-commit-config.yaml
… .prettierrc
… .vscode/extensions.json
… biome.json
… package.json
… static/app/components/dynamicSampling/investigationRule.spec.tsx
… static/app/components/events/eventErrors.tsx
… static/app/components/events/eventStatisticalDetector/eventThroughput.tsx
… static/app/components/events/interfaces/crashContent/exception/actionableItemsUtils.tsx
… static/app/components/events/interfaces/crashContent/exception/relatedExceptions.spec.tsx
… static/app/components/events/interfaces/exceptionV2.tsx
… static/app/components/events/interfaces/frame/line/native.tsx
… static/app/components/events/interfaces/frame/stacktraceLink.tsx
… static/app/components/events/interfaces/frame/stacktraceLinkModal.tsx
… static/app/components/events/interfaces/performance/spanEvidenceKeyValueList.tsx
… static/app/components/events/interfaces/performance/utils.tsx
… static/app/components/events/interfaces/spans/spanTreeModel.tsx
… static/app/components/events/interfaces/spans/utils.tsx
… static/app/components/events/interfaces/stackTrace.tsx
… static/app/components/events/interfaces/threads.tsx
… static/app/components/events/traceEventDataSection.tsx
… static/app/components/events/viewHierarchy/wireframe.tsx
… static/app/components/feedback/useMutateFeedback.tsx
… static/app/components/forms/controls/selectAsyncControl.tsx
… static/app/components/forms/fields/projectMapperField.tsx
… static/app/components/forms/fields/selectAsyncField.tsx
… static/app/components/forms/jsonForm.spec.tsx
… static/app/components/forms/model.tsx
… static/app/components/globalSdkUpdateAlert.tsx
… static/app/components/globalSelectionLink.tsx