@anonrig 2024-01-18
> Is there a reason we did not merge this? (Or the other one)

I couldn't find anybody to land this and later help me track the health of the release, in case there is a bug.